add action for Like
