+++ src/components/redux/actions/Post.js
@@ -1,7 +1,6 @@
 import * as types from '../const/actionTypes/Post';
 import request from 'superagent';
 import url from 'components/const/StaticData';
-import store from 'components/redux/store';
 
 const requestPost = (id) => ({
   type: types.POST_FETCH_REQUEST,
@@ -27,27 +26,5 @@ export function fetchPost(id) {
               err ? dispatch(requestError())
                   : dispatch(requestSuccess(response.body));
             });
-  };
-}
-
-const likePlus = (item) => ({
-  type: types.POST_INC_LIKE,
-  item
-});
-
-const postLike = (id, like) =>  (
-  request.put(`${url}/post`)
-  .set('Content-Type', 'application/json')
-  .accept('application/json')
-  .send(`{"id":${id},"like":${like}}`)
-  .end(null)
-);
-
-export function incLike() {
-  return (dispatch) => {
-    const item = Object.assign({}, store.getState().post.entry);
-    item.metadata.like += 1;
-    postLike(item.id, item.metadata.like);
-    dispatch(likePlus(item));
   };
 }
