+++ src/components/redux/const/actionTypes/Posts.js
@@ -1,4 +1,4 @@
 export const POSTS_FETCH_REQUEST = 'POSTS_FETCH_REQUEST';
 export const POSTS_FETCH_SUCCESS = 'POSTS_FETCH_SUCCESS';
 export const POSTS_FETCH_ERROR = 'POSTS_FETCH_ERROR';
-export const POST_INC_LIKE = 'POST_INC_LIKE';
+export const POSTS_INC_LIKE = 'POSTS_INC_LIKE';
